+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/dto/PersonDTO.java
@@ -56,4 +56,8 @@ public class PersonDTO {
     public boolean isParent() {
         return parent;
     }
+
+    public void setParent(boolean b) {
+        this.parent = b;
+    }
 }